:lipstick: feat: Estilizaçao da página de pizzas e botões

+++ modulo6/rodada-case2/projeto-ambulnz-backend/src/database/migrations/Migrations.ts
@@ -37,7 +37,7 @@ class Migrations extends BaseDatabase {
     
     CREATE TABLE IF NOT EXISTS ${PizzaDatabase.TABLE_PIZZAS} (
         name VARCHAR(255) PRIMARY KEY,
-        price DECIMAL(3,2) NOT NULL,
+        price float NOT NULL,
         image_url VARCHAR(255) NOT NULL
     );
